requirements.txt
streamlit
python-dotenv
openai>=1.0.0
tiktoken
PyPDF2
python-pptx
tqdm
pdfplumber
numpy
pandas
crewai
crewai_tools
langchain-openai
pinecone<7.0.0,>=6.0.0
langchain
